Refactoring of the directories. - Separates the ea-nas spesific code to ea_nas directory - leaves src free to be occupied by common code for both pattenNets and ea_nas
