+++ Dockerfile
@@ -13,7 +13,7 @@ RUN pip install setproctitle
 
 # Copy over codefiles:
 VOLUME /src
-COPY ./datasets /src/datasets
+COPY src/training /src/datasets
 COPY ./firebase /src/firebase
 COPY ./resources /src/resources
 COPY ./src /src/src
@@ -25,6 +25,6 @@ WORKDIR /src
 ENV LANG="C.UTF-8"
 ENV EA_NAS_UPLOAD_TO_FIREBASE="1"
 
-# Ready to run: 
+# Ready to run:
 # RUN python tests.py
 CMD python -u start.py ./datasets/cifar10-telenor001.json
