+++ configurations/test_local_cpu.json
@@ -1,0 +1,30 @@
+{
+  "dataset": "CIFAR-10",
+  "trainingFilepath": "./src/training/cifar10.py",
+  "input": [
+    32,
+    32,
+    3
+  ],
+  "classes": 10,
+  "batch size": 300,
+  "epochs": 0,
+  "population size": 1,
+  "generations": 1,
+  "run id": "simulation01",
+  "keep all results": false,
+  "servers": [{
+    "name": "COMPUTE",
+    "type": "local",
+    "cwd": "/Users/magnus/projects/python/nas",
+    "python": "/Users/magnus/projects/python/nas/venv/bin/python",
+    "address": "localhost",
+    "devices": [{
+      "device_str": "/CPU:0",
+      "allow gpu memory growth": true,
+      "memory per process": 1,
+      "concurrency": 1
+    }]
+  }]
+}
+
