+++ configurations/test_remote.json
@@ -1,6 +1,6 @@
 {
   "dataset": "CIFAR-10",
-  "trainingFilepath": "datasets/cifar10.py",
+  "trainingFilepath": "./src/training/cifar10.py",
   "input": [
     32,
     32,
