--- datasets/cifar-10-mp.json
@@ -1,33 +1,0 @@
-{
-  "dataset": "CIFAR-10",
-  "trainingFilepath": "datasets/cifar10.py",
-  "input": [
-    32,
-    32,
-    3
-  ],
-  "classes": 10,
-  "batch size": 40,
-  "epochs": 0.50,
-  "population size": 4,
-  "generations": 10,
-  "run id": "8x2",
-  "keep all results": false,
-  "servers": [{
-    "name": "ANAKIN01",
-    "type": "local",
-    "cwd": "/work/magnuspw/ea-nas",
-    "address": "anakin01.idi.ntnu.no",
-    "devices": [{
-      "device_str": "/GPU:0",
-      "allow gpu memory growth": true,
-      "memory per process": 0.5,
-      "concurrency": 2
-    }, {
-      "device_str": "/GPU:1",
-      "allow gpu memory growth": true,
-      "memory per process": 0.5,
-      "concurrency": 2
-    }]
-  }]
-}
