+++ src/pattern_nets/operations.py
