+++ src/training/mnist.py
